Add files via upload
This file contains the vector representation in LIWC categories of dreams of 20 patients along with their analyses by several professional analysts. The last analysis is ChatGPT. This can be seen also in the column isChatGPT. Dreams are noted by letters D1, D2, and Analyses by A1, A2, A3, etc. The column isDream is 0 when the row represents an analysis, and is 1 when the row represents a dream.